test(app): strengthen controller revision properties
